Restore OrderDetailsDataSource file so it can be refactored in a separate PR
